fix: resolve circular dependancies
