+++ alma/upgrade/upgrade-3.0.0.php
@@ -26,6 +26,7 @@ use Alma\PrestaShop\Builders\Helpers\ApiHelperBuilder;
 use Alma\PrestaShop\Forms\InpageAdminFormBuilder;
 use Alma\PrestaShop\Helpers\ConstantsHelper;
 use Alma\PrestaShop\Helpers\SettingsHelper;
+use Alma\PrestaShop\Services\InsuranceService;
 
 if (!defined('_PS_VERSION_')) {
     exit;
@@ -38,7 +39,9 @@ function upgrade_module_3_0_0($module)
     try {
         $apiHelperBuilder = new ApiHelperBuilder();
         $apiHelper = $apiHelperBuilder->getInstance();
-        $apiHelper->getMerchant();
+        $merchant = $apiHelper->getMerchant();
+        $insuranceService = new InsuranceService();
+        $insuranceService->installIfCompatible($merchant);
     } catch (\Exception $e) {
     }
